+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [9.13.0] - 2026-08-20
 
 ### Added